pyproject.toml
[project]
name = "graphweaver-agent"
version = "1.0.0"
description = "Claude-powered autonomous FK discovery agent with MCP + LangChain"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # Database
    "psycopg2-binary>=2.9.0",
    "neo4j>=5.0.0",
    
    # LangChain + Claude
    "langchain-anthropic>=0.3.10",
    "langchain-core>=0.3.30",
    "langgraph>=0.2.60",
    
    # MCP
    "mcp>=1.0.0",
    
    # Data validation
    "pydantic>=2.10.0",
    
    # HTTP requests (for Marquez)
    "requests>=2.31.0",
    
    # YAML parsing
    "pyyaml>=6.0.0",
    
    # Text Embeddings
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    
    # Numerical operations
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
graphweaver-agent = "agent:main"
graphweaver-mcp = "mcp_servers.graphweaver_mcp:mcp.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/graphweaver_agent"]